% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairGenes.R
\name{CalStabRvsPair}
\alias{CalStabRvsPair}
\title{CalStabRvsPair}
\usage{
CalStabRvsPair(FindPairFrame, ControlFrame, TreatFrame)
}
\arguments{
\item{FindPairFrame}{The Pair Frame.}

\item{ControlFrame}{Expression matrix of control group.}

\item{TreatFrame}{Expression matrix of treat group.}
}
\value{
Sort out the sample-gene pair matrix by reversing the gene pair matrix.
}
\description{
CalStabRvsPair
}
\examples{
CalStabRvsPair(FindPairFrame,ControlFrame,TreatFrame)
}
