% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairTime.R
\name{testRunTime}
\alias{testRunTime}
\title{testRunTime}
\usage{
testRunTime(ControlFrame,TreatFrame,HighPercent=0.9,ncore=1)
}
\arguments{
\item{ControlFrame}{Expression matrix of control group.}

\item{TreatFrame}{Expression matrix of treat group.}

\item{HighPercent}{Percentage of samples with higher expression of Gene i than Gene j in total samples.}

\item{LowPercent}{Percentage of samples with lower expression of Gene i than Gene j in total samples.}
}
\value{
The approximate time required for this computer to return the data.
}
\description{
testRunTime
}
\examples{
testRunTime(ControlFrame,TreatFrame,HighPercent=0.9,ncore=1)
}
